+++ src/NATS.Client.Core/NatsConnection.cs
@@ -28,7 +28,6 @@ public partial class NatsConnection : IAsyncDisposable, INatsConnection
     private readonly object _gate = new object();
     private readonly WriterState _writerState;
     private readonly ChannelWriter<ICommand> _commandWriter;
-    internal readonly SubscriptionManager _subscriptionManager;
     private readonly ILogger<NatsConnection> _logger;
     private readonly ObjectPool _pool;
     private readonly CancellationTimerPool _cancellationTimerPool;
@@ -69,7 +68,7 @@ public partial class NatsConnection : IAsyncDisposable, INatsConnection
         _writerState = new WriterState(options);
         _commandWriter = _writerState.CommandBuffer.Writer;
         InboxPrefix = $"{options.InboxPrefix}.{Guid.NewGuid():n}.";
-        _subscriptionManager = new SubscriptionManager(this, InboxPrefix);
+        SubscriptionManager = new SubscriptionManager(this, InboxPrefix);
         _logger = options.LoggerFactory.CreateLogger<NatsConnection>();
         _clientOptions = new ClientOptions(Options);
         HeaderParser = new HeaderParser(options.HeaderEncoding);
@@ -81,6 +80,8 @@ public partial class NatsConnection : IAsyncDisposable, INatsConnection
     public event EventHandler<string>? ConnectionOpened;
 
     public event EventHandler<string>? ReconnectFailed;
+
+    internal SubscriptionManager SubscriptionManager { get; }
 
     public NatsOptions Options { get; }
 
@@ -143,7 +144,7 @@ public partial class NatsConnection : IAsyncDisposable, INatsConnection
                 item.SetCanceled();
             }
 
-            await _subscriptionManager.DisposeAsync().ConfigureAwait(false);
+            await SubscriptionManager.DisposeAsync().ConfigureAwait(false);
             _waitForOpenConnection.TrySetCanceled();
             _disposedCancellationTokenSource.Cancel();
         }
@@ -167,7 +168,7 @@ public partial class NatsConnection : IAsyncDisposable, INatsConnection
 
     internal ValueTask PublishToClientHandlersAsync(string subject, string? replyTo, int sid, in ReadOnlySequence<byte>? headersBuffer, in ReadOnlySequence<byte> payloadBuffer)
     {
-        return _subscriptionManager.PublishToClientHandlersAsync(subject, replyTo, sid, headersBuffer, payloadBuffer);
+        return SubscriptionManager.PublishToClientHandlersAsync(subject, replyTo, sid, headersBuffer, payloadBuffer);
     }
 
     internal void ResetPongCount()
@@ -383,14 +384,6 @@ public partial class NatsConnection : IAsyncDisposable, INatsConnection
             _writerState.PriorityCommands.Add(connectCommand);
             _writerState.PriorityCommands.Add(PingCommand.Create(_pool, GetCancellationTimer(CancellationToken.None)));
 
-            // if (reconnect)
-            // {
-            //     // Add SUBSCRIBE command to priority lane
-            //     var subscribeCommand =
-            //         AsyncSubscribeBatchCommand.Create(_pool, GetCancellationTimer(CancellationToken.None), _subscriptionManager.GetExistingSubscriptions().ToArray());
-            //     _writerState.PriorityCommands.Add(subscribeCommand);
-            // }
-
             // create the socket writer
             _socketWriter = new NatsPipeliningWriteProtocolProcessor(_socket!, _writerState, _pool, Counter);
 
@@ -400,7 +393,8 @@ public partial class NatsConnection : IAsyncDisposable, INatsConnection
             // receive COMMAND response (PONG or ERROR)
             await waitForPongOrErrorSignal.Task.ConfigureAwait(false);
 
-            await _subscriptionManager.ReconnectAsync(_disposedCancellationTokenSource.Token).ConfigureAwait(false);
+            // Reestablish subscriptions and consumers
+            await SubscriptionManager.ReconnectAsync(_disposedCancellationTokenSource.Token).ConfigureAwait(false);
         }
         catch (Exception)
         {
