+++ src/NATS.Client.Core/NatsSub.cs
@@ -206,25 +206,6 @@ public abstract class NatsSubBase
     }
 }
 
-public interface INatsSub : IAsyncDisposable
-{
-    ChannelReader<NatsMsg> Msgs { get; }
-
-    /// <summary>
-    /// The subject name to subscribe to.
-    /// </summary>
-    string Subject { get; }
-
-    /// <summary>
-    /// If specified, the subscriber will join this queue group. Subscribers with the same queue group name,
-    /// become a queue group, and only one randomly chosen subscriber of the queue group will
-    /// consume a message each time a message is received by the queue group.
-    /// </summary>
-    string? QueueGroup { get; }
-
-    public ValueTask UnsubscribeAsync();
-}
-
 public sealed class NatsSub : NatsSubBase, INatsSub
 {
     private static readonly BoundedChannelOptions DefaultChannelOptions =
@@ -286,25 +267,6 @@ public sealed class NatsSub : NatsSubBase, INatsSub
     }
 
     protected override void TryComplete() => _msgs.Writer.TryComplete();
-}
-
-public interface INatsSub<T> : IAsyncDisposable
-{
-    ChannelReader<NatsMsg<T?>> Msgs { get; }
-
-    /// <summary>
-    /// The subject name to subscribe to.
-    /// </summary>
-    string Subject { get; }
-
-    /// <summary>
-    /// If specified, the subscriber will join this queue group. Subscribers with the same queue group name,
-    /// become a queue group, and only one randomly chosen subscriber of the queue group will
-    /// consume a message each time a message is received by the queue group.
-    /// </summary>
-    string? QueueGroup { get; }
-
-    public ValueTask UnsubscribeAsync();
 }
 
 public sealed class NatsSub<T> : NatsSubBase, INatsSub<T>
